requirements.txt
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
pyzmq==14.3.1
requests==2.4.3
rethinkdb==1.15.0-1
tornado==4.0.2
wsgiref==0.1.2
arrow==0.4.2
docker_links_python==0.1.0
